Product service and implementation

+++ src/Core/E-Ticaret Project.Application/DTOs/ProductDtos/ProductCreateDto.cs
@@ -6,7 +6,7 @@ public record ProductCreateDto(
     string Tittle,
     string? Description,
     decimal Price,
-    decimal? Discount,
+    int? Discount,
     decimal? Rating,
     int Stock,
     List<IFormFile>? image,
